requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.1
Django==4.1.7
django-crispy-forms==2.0
django-jwt-extended==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
